+++ client/src/notebooks/components/StartNotebookServerOptions.tsx
@@ -33,7 +33,7 @@ import {
 } from "reactstrap";
 
 import styles from "./StartNotebookServerOptions.module.scss";
-import { ResourcePoolPicker } from "./ResourcePoolPicker";
+import { ResourcePoolPicker } from "./SessionClassSelector";
 import { useServerOptionsQuery } from "../../features/session/sessionApi";
 import { Loader } from "../../components/Loader";
 import { RootStateOrAny, useDispatch, useSelector } from "react-redux";
